chore

+++ agent/scrcpy/src/main/java/com/yqhp/agent/scrcpy/message/Position.java
@@ -9,6 +9,6 @@ import lombok.Data;
 public class Position {
     private int x;
     private int y;
-    private int width;
-    private int height;
+    private short width;
+    private short height;
 }
